fix: ambiguity in feature model semantics
The translation of feature models to propositional formulas was slightly
ambiguous.
Within the translation of or-groups, the formula
\bigvee_{1 \leq i \leq n} C_i \pimplies P
could be read as
\bigvee_{1 \leq i \leq n} (C_i \pimplies P)
or
(\bigvee_{1 \leq i \leq n} C_i) \pimplies P
where only the first one is correct as defined by the binding
precedences.
Still it was confusing.
This commit also adapts another rule for consistency.